Moving tests back into IntrospectionTest
