+++ tests/Functional/IntrospectionTest.php
@@ -452,4 +452,76 @@ class IntrospectionTest extends TestCase
         ], $result);
     }
 
+    /**
+     * Test to check that we can introspect on a scalar, which does *not*
+     * have a name which clashes with a global PHP function or class
+     */
+    public function testCanIntrospectOnANonClashingScalar()
+    {
+        $schema = '
+        scalar Postcode
+        type Query {
+            hello: Postcode
+        }
+        ';
+
+
+        $schema = buildSchema($schema);
+
+        $query = '
+        query IntrospectionDroidDescriptionQuery {
+          __type(name: "Postcode") {
+            name
+          }
+        }
+        ';
+
+        $result = graphql($schema, $query);
+
+        $this->assertArrayNotHasKey('errors', $result);
+        $this->assertEquals([
+            'data' => [
+                '__type' => [
+                    'name'        => 'Postcode',
+                ],
+            ],
+        ], $result);
+    }
+
+    /**
+     * Test to check that we can introspect on a scalar, which *does*
+     * have a name which clashes with a global PHP function or class
+     */
+    public function testCanIntrospectOnScalarWithClashingName()
+    {
+        $schema = '
+        scalar Date
+        type Query {
+            hello: Date
+        }
+        ';
+
+
+        $schema = buildSchema($schema);
+
+        $query = '
+        query IntrospectionDroidDescriptionQuery {
+          __type(name: "Date") {
+            name
+          }
+        }
+        ';
+
+        $result = graphql($schema, $query);
+
+        $this->assertArrayNotHasKey('errors', $result);
+        $this->assertEquals([
+            'data' => [
+                '__type' => [
+                    'name'        => 'Date',
+                ],
+            ],
+        ], $result);
+    }
+
 }
